Correct the repo hash caculation algorithm.
If the caller and callee satisfies the following conditions, the error of undefined reference to the callee is given.
    1. Caller is pruned and deleted during the RepoPruning pass.
    2. Callee isn’t pruned, has noinline attribute and may be discarded if it is not used.

The reason is that the callee is removed during the Dead Global Elimination pass, which is after the repo pruning pass.

The current hash calculation algorithm doesn’t accumulate the callee hash value to the caller if the callee has noinline attribute.
To solve the issue, the caller always accumulates the callee’s hash value if callee may be discarded if it is not used.

Fix for <https://github.com/SNSystems/llvm-project-prepo/issues/76>.