Fixed a pruning issue for an alias case.
Current pruning implementation for the alias exists an issue for the following case.
Assuming: 1) the aliasee of alias 'A' is alias 'B',
          2) the aliasee of alias 'B' is global object 'C',
          3) the global object 'C' is pruned.
The order in the alias list is 'B' and then 'A'. Since the global object 'C' is prund,
the alias 'B' is changed from definition to declaration. Then, since 'B' is changed to
declaration, the base object of the alias 'A' will be changed from 'C' to 'B'. This
will result in the backend error.

To solve the issue, the aliases' pruning is separated into two steps.
First, check whether a GlobalAlias references a pruned GO. If yes, record it.
Second, replace all recorded aliases by the declaration.

Fix for <https://github.com/SNSystems/llvm-project-prepo/issues/81>.