Place ELFState<> on the heap with naked new.
Use of "naked" new to instantiate objects is, in general, a bad idea
with an RAII helper such as unique_ptr<> normally preferred. However,
the various data structures maintained with ELFState<> take measurable
time to tear-down (I measured up to 30% of repo2obj's overall
runtime). The simplest way to avoid this overhead is to simply leak
the object and let the OS release the memory when the process exists.