Use unordered_map for the symbols and string storage. (#87)
pstore makes it extremely fast to compare strings for equality:
they're stored in a set meaning that two strings with the same address
are equal. Lexicographical comparison OTOH requires the strings to be
loaded and compared code-unit by code-unit.

Previously the code used map<> which relies on lexiographical
order. This change switches to unordered_map<> so that (once hashed)
we only need equality.