setSymbolShadow() spin-lock did not bump its iteration count.
We have two spin-lock implementations: one is the SpinLock class and
the other is the code in setSymbolShadow() that protects the symbol
shadow pointers. The latter did not increment its iteration count
meaning that it would never "back-off" by calling this_thread::yield()
in the event of contention.

The number of spins before we call this_thread::yield is now a
constant shared between the two variants.