Minor improvement to layout dump. Fix layout for overridden symbols.
Dump the input ordinal for each section-info object in the layout. The
path of the file that corresponds to this would be more useful, but
this will do in the short term.

The previous version asserted if a symbol body's ordinal did not match
that of the file being processed: this fixes that behaviour.