Performance Improvement for the hash algorithm.
When calculating the GO's hash, if GO's dependencies is a function, which will not be inlined and not be discarded if it is not used, the function hash will not be accumulated to the GO's hash. We could move the check to the earlier stage when building the GO's dependencies. If a function is satisfied the above conditions, it will not be added into the dependencies. It will reduce the size of GO's dependencies.

The llvm-project (commit: a822ec75) is used to measure the performance improvement.

The backend compilation time of the repo compiler (master: d93faf2) is 17166.0s. After this commit, the backend compilation time is 15604.4s. The backend compilation time is reduced by 1,561.6s (9.1% improvement).