The problem is caused by the incorrect hash calculation. When we calculate the initial digest for each global object (GO), we also collect the GO's ‘Contributions’ and ‘Dependencies’. The GO’s ‘Contributions’ and ‘Dependencies’ represent the different directions during the hash calculation. The detailed explanation of the repo hash calculation is given in https://github.com/SNSystems/llvm-project-prepo/wiki/The-RepoMetadataGeneration-and-RepoPruning-Passes.
The case that shows the problem is:

int V1 = 1;
int V2 = 0;
int func() {
  V2 = V2 + V1;
  return V2;
}

After the initial hash calculation, the following information is collected.
‘V1’: Contributions: [ ], Dependencies: [ ].
‘V2’: Contributions: [ ], Dependencies: [ ].
‘func’: Contributions: [ ‘V2’], Dependencies: [‘V1’].

The current hash algorithm only accumulates the `func` hash to the ‘V2’ hash, which is not enough. If ‘V1’ initial value changes, the ‘V2’ value will change. Therefore, the ‘V1’ hash needs to be accumulated to the ‘V2’ hash. The correct informatiopn for 'func' should be:

‘func’: Contributions: [ 'V1', ‘V2’], Dependencies: [ ].

By converting the ‘Contributions’ to ‘Dependencies’, it will fix the issue because the hash algorithm will only consider the GO’s dependencies, which are in the same direction when calculating the Go’s hash.

The commit includes the following changes:
1) Rename ‘ContributedToGVs’ to ‘Contributions’ and make sure there is no intersection between the GO’s Contribution and its Dependencies.
2) Convert the 'Contributions' to 'Dependencies' to get a single direction graph, which is used by the hash calculation.

Fix for <https://github.com/SNSystems/llvm-project-prepo/issues/110>.