The problem is caused by the incorrect hash calculation. When we calculate the initial digest for each global object (GO), we also collect the GO's ‘Contributions’ and ‘Dependencies’. The GO’s ‘Contributions’ and ‘Dependencies’ represent the different directions during the hash calculation. If GO’s digest contributes to other GOs, we will put all the other GOs into the GO’s ‘Contributions’. However, the ‘Dependencies’ of the GO are those objects whose digest value changes might affect the GO’s digest value. Both ‘Contributions’ and ‘Dependencies’ are used by the hash calculation pass. The Dependencies of a GO’s contribution should be used to calculate the GO’s hash value, but they did not. This results in the incorrect hash.
The case that shows the problem is:

int V1 = 1;
int V2 = 0;
int func() {
  V2 = V2 + V1;
  return V2;
}

After the initial hash calculation, the following information is collected.
‘V1’: Contributions: [ ], Dependencies: [ ].
‘V2’: Contributions: [ ], Dependencies: [ ].
‘func’: Contributions: [ ‘V2’], Dependencies: [‘V2’, ‘V1’].

The current hash algorithm only accumulates the `func` hash to the ‘V2’ hash, which is not enough. If ‘V1’ initial value changes, the ‘V2’ value will change. Therefore, the ‘V1’ hash needs to be accumulated to the ‘V2’ hash.

I think that by converting the ‘Contribution’ to ‘Dependencies, it will fix it because the hash algorithm will only consider the GO’s dependencies, which are in the same direction when calculating the Go’s hash.

The commit includes the following changes:
1) Rename ‘ContributedToGVs’ to ‘Contributions’.
2) Convert the 'Contributions' to 'Dependencies' to get a single direction graph, which is used by the hash calculation.