The 'names' argument is positional. Add short switches.
In starting to use the tool in anger I found that I much preferred the
switches employed by the original python implementation. That is, the
'names' switch was simply positional arguments listed after the ticket
file, and I had single letter aliases for the --digest-only and
--comma switches.

Extend the test to cover these new switches. Don't set the REPOFILE
environment variable where we're using the --repo switch. Give the
FileCheck prefixes meaningful names.