Unify the code for ticket files and finalise their contents.
There is now a single module which understands the format of ticket
files and can read and write them. Extend the file contents so that it
includes a version number, CRC, and owning database ID.

Extend the tools so that the owning database ID is checked and add
tests to ensure that an error is issued if a mismatch is found.