Bring the prepo x86 object writer into line with the ELF version.
Whilst compiling with a fairly recent version of GCC, I noticed a
series of warnings being issued for X86RepoObjectWriter.cpp that were
not produced when compiling X86ELFObjectWriter.cpp. The repo code was
based on an older version of the ELF output code which has clearly
moved on since that fork was taken.

This commit brings the Repo implementation into line with the ELF
equivalent and resolves those compiler warnings.