Recording the compilation&entry-index for linked-definitions.
Reflecting changes in the pstore API to record the compilation digest
and index of the definition referenced by an entry in a
linked-definition section.

Fix a bug caused by calling code failing to keep the pointer returned
from pstore alive which provoked a change to the get_sstring_view()
helper function.