Add the repo-create-ticket tool. (#126)
repo-create-ticket is intended as a companion to repo-ticket-dump. The
latter will tell you the compilation digest associated with an
existing ticket file. repo-create-ticket will create a new ticket file
given an existing compilation digest.

With this new tool, the pstore-import tool, and repository contents
described by a JSON file (probably written by pstore-export) it is
possible to construct a repository with specific characteristics for
use in testing tools such as repo2obj or rld.