Fix for issue #151. Map PHDRs to the process address space.
Yet another reworking of how we generate ELF program headers. This
time, knowledge of the overlap of the ELF header/program header is
available in flattenSegments() so that the ossets are correctly
propogated to output sections and contributions. The number of ELF
program headers is almost computed at this point.