Remove the following warning: clang: warning: argument unused during compilation: '-stdlib=libc++'.
On systems where libc++ is provided but is not the default, Clang provides a flag called
-stdlib= that can be used to decide which standard library is used. Using -stdlib=libc++
will select libc++. However, we use a custom built libc++ library.  Clang provides a way
to disable the default behavior for finding the standard library and to override it with
custom paths. This can be done with:
“-nostdinc++ -nostdlib++ -isystem <install>/include/c++/v1 -L <install>/lib -lc++”.

This commit includes:
1. Fixed the warning.
2. Since the unwind table section isn't supported on repo target yet.
add the -fno-unwind-tables to the compile flags.