Improve duplicate symbol reporting.
When we are raising an error about a duplicate symbol, we need to tell
the user about the existing location as well as the new. The
defineSymbol()/referenceSymbol() code now returns a symbol pointer and
a bool indicating success or otherwise instead of a symbol pointer
where null indicates failure.